NP-14 - Update Name Phonetics to require OpenMRS 2.4+ in the main branch and increment to release 2.0.0-SNAPSHOT.
