--- api-2.x/pom.xml
@@ -1,80 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<parent>
-		<groupId>org.openmrs.module</groupId>
-		<artifactId>namephonetics</artifactId>
-		<version>1.19.0-SNAPSHOT</version>
-	</parent>
-	<groupId>org.openmrs.module</groupId>
-	<artifactId>namephonetics-api-2.x</artifactId>
-	<packaging>jar</packaging>
-	<name>Name Phonetics API for the 2.x platform</name>
-	<description>API project for Name Phonetics</description>
-
-	<dependencies>
-		<!-- 
-		Add other dependencies from lib:
-		<dependency>
-			<groupId>org.other.library</groupId>
-			<artifactId>library-name</artifactId>
-		</dependency>
-		-->
-		
-		<dependency>
-			<groupId>org.openmrs.module</groupId>
-			<artifactId>namephonetics-api</artifactId>
-			<version>${project.parent.version}</version>
-		</dependency>
-	
-		<dependency>
-			<groupId>commons-codec</groupId>
-			<artifactId>commons-codec</artifactId>
-			<version>1.4</version>
-		</dependency>
-						
-		<dependency>
-		    <groupId>org.openmrs.api</groupId>
-		    <artifactId>openmrs-api</artifactId>
-		    <version>${openMRSVersion}</version>
-		    <type>jar</type>
-		    <scope>provided</scope>
-		</dependency>
-		 
-		<dependency>
-		    <groupId>org.openmrs.api</groupId>
-		    <artifactId>openmrs-api</artifactId>
-		    <version>${openMRSVersion}</version>
-		    <type>test-jar</type>
-		    <scope>test</scope>
-		</dependency>
-		 
-		<dependency>
-		    <groupId>org.openmrs.web</groupId>
-		    <artifactId>openmrs-web</artifactId>
-		    <version>${openMRSVersion}</version>
-		    <type>jar</type>
-		    <scope>provided</scope>
-		</dependency>
-		 
-		<dependency>
-		    <groupId>org.openmrs.web</groupId>
-		    <artifactId>openmrs-web</artifactId>
-		    <version>${openMRSVersion}</version>
-		    <type>test-jar</type>
-		    <scope>test</scope>
-		</dependency>
-		 
-		<dependency>
-		    <groupId>org.openmrs.test</groupId>
-		    <artifactId>openmrs-test</artifactId>
-		    <version>${openMRSVersion}</version>
-		    <type>pom</type>
-		    <scope>test</scope>
-		</dependency>
-	</dependencies>
-	
-	<properties>
-		<openMRSVersion>2.0.0</openMRSVersion>
-	</properties>
-
-</project>
