--- api-tests/pom.xml
@@ -1,78 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<parent>
-		<groupId>org.openmrs.module</groupId>
-		<artifactId>namephonetics</artifactId>
-		<version>1.19.0-SNAPSHOT</version>
-	</parent>
-	<groupId>org.openmrs.module</groupId>
-	<artifactId>namephonetics-api-tests</artifactId>
-	<packaging>jar</packaging>
-	<name>Name Phonetics API Tests</name>
-	<description>API Test project for Name Phonetics</description>
-
-	<dependencies>
-
-		<dependency>
-			<groupId>commons-codec</groupId>
-			<artifactId>commons-codec</artifactId>
-			<version>1.4</version>
-		</dependency>
-						
-		<dependency>
-		    <groupId>org.openmrs.api</groupId>
-		    <artifactId>openmrs-api</artifactId>
-		    <version>${openMRSVersion}</version>
-		    <type>jar</type>
-		    <scope>provided</scope>
-		</dependency>
-		 
-		<dependency>
-		    <groupId>org.openmrs.api</groupId>
-		    <artifactId>openmrs-api</artifactId>
-		    <version>${openMRSVersion}</version>
-		    <type>test-jar</type>
-		    <scope>test</scope>
-		</dependency>
-		 
-		<dependency>
-		    <groupId>org.openmrs.web</groupId>
-		    <artifactId>openmrs-web</artifactId>
-		    <version>${openMRSVersion}</version>
-		    <type>jar</type>
-		    <scope>provided</scope>
-		</dependency>
-		 
-		<dependency>
-		    <groupId>org.openmrs.web</groupId>
-		    <artifactId>openmrs-web</artifactId>
-		    <version>${openMRSVersion}</version>
-		    <type>test-jar</type>
-		    <scope>test</scope>
-		</dependency>
-		 
-		<dependency>
-		    <groupId>org.openmrs.test</groupId>
-		    <artifactId>openmrs-test</artifactId>
-		    <version>${openMRSVersion}</version>
-		    <type>pom</type>
-		    <scope>test</scope>
-		</dependency>
-
-		<dependency>
-			<groupId>${project.groupId}</groupId>
-			<artifactId>${project.parent.artifactId}-api</artifactId>
-			<version>${project.version}</version>
-			<scope>provided</scope>
-		</dependency>
-
-		<dependency>
-			<groupId>${project.groupId}</groupId>
-			<artifactId>${project.parent.artifactId}-api-1.x</artifactId>
-			<version>${project.version}</version>
-			<scope>provided</scope>
-		</dependency>
-
-	</dependencies>
-
-</project>
