+++ api/pom.xml
@@ -3,68 +3,12 @@
 	<parent>
 		<groupId>org.openmrs.module</groupId>
 		<artifactId>namephonetics</artifactId>
-		<version>1.19.0-SNAPSHOT</version>
+		<version>2.0.0-SNAPSHOT</version>
 	</parent>
 	<groupId>org.openmrs.module</groupId>
 	<artifactId>namephonetics-api</artifactId>
 	<packaging>jar</packaging>
 	<name>Name Phonetics API</name>
 	<description>API project for Name Phonetics</description>
-
-	<dependencies>
-		<!-- 
-		Add other dependencies from lib:
-		<dependency>
-			<groupId>org.other.library</groupId>
-			<artifactId>library-name</artifactId>
-		</dependency>
-		-->
-		
-		<dependency>
-			<groupId>commons-codec</groupId>
-			<artifactId>commons-codec</artifactId>
-			<version>1.4</version>
-		</dependency>
-						
-		<dependency>
-		    <groupId>org.openmrs.api</groupId>
-		    <artifactId>openmrs-api</artifactId>
-		    <version>${openMRSVersion}</version>
-		    <type>jar</type>
-		    <scope>provided</scope>
-		</dependency>
-		 
-		<dependency>
-		    <groupId>org.openmrs.api</groupId>
-		    <artifactId>openmrs-api</artifactId>
-		    <version>${openMRSVersion}</version>
-		    <type>test-jar</type>
-		    <scope>test</scope>
-		</dependency>
-		 
-		<dependency>
-		    <groupId>org.openmrs.web</groupId>
-		    <artifactId>openmrs-web</artifactId>
-		    <version>${openMRSVersion}</version>
-		    <type>jar</type>
-		    <scope>provided</scope>
-		</dependency>
-		 
-		<dependency>
-		    <groupId>org.openmrs.web</groupId>
-		    <artifactId>openmrs-web</artifactId>
-		    <version>${openMRSVersion}</version>
-		    <type>test-jar</type>
-		    <scope>test</scope>
-		</dependency>
-		 
-		<dependency>
-		    <groupId>org.openmrs.test</groupId>
-		    <artifactId>openmrs-test</artifactId>
-		    <version>${openMRSVersion}</version>
-		    <type>pom</type>
-		    <scope>test</scope>
-		</dependency>
-	</dependencies>
 
 </project>
