+++ omod/pom.xml
@@ -3,7 +3,7 @@
 	<parent>
 		<groupId>org.openmrs.module</groupId>
 		<artifactId>namephonetics</artifactId>
-		<version>1.19.0-SNAPSHOT</version>
+		<version>2.0.0-SNAPSHOT</version>
 	</parent>
 	<groupId>org.openmrs.module</groupId>
 	<artifactId>namephonetics-omod</artifactId>
@@ -12,60 +12,36 @@
 	<description>OpenMRS module project for Name Phonetics</description>
 
 	<dependencies>
-	<dependency>
-		<groupId>org.openmrs.module</groupId>
-		<artifactId>namephonetics-api</artifactId>
-		<version>${project.parent.version}</version>
-	</dependency>
-	<dependency>
-		<groupId>org.openmrs.module</groupId>
-		<artifactId>namephonetics-api-1.x</artifactId>
-		<version>${project.parent.version}</version>
-	</dependency>
-	<dependency>
-		<groupId>org.openmrs.module</groupId>
-		<artifactId>namephonetics-api-2.x</artifactId>
-		<version>${project.parent.version}</version>
-	</dependency>
-	<dependency>
-	    <groupId>org.openmrs.api</groupId>
-	    <artifactId>openmrs-api</artifactId>
-	    <version>${openMRSVersion}</version>
-	    <type>jar</type>
-	    <scope>provided</scope>
-	</dependency>
-	 
-	<dependency>
-	    <groupId>org.openmrs.api</groupId>
-	    <artifactId>openmrs-api</artifactId>
-	    <version>${openMRSVersion}</version>
-	    <type>test-jar</type>
-	    <scope>test</scope>
-	</dependency>
-	 
-	<dependency>
-	    <groupId>org.openmrs.web</groupId>
-	    <artifactId>openmrs-web</artifactId>
-	    <version>${openMRSVersion}</version>
-	    <type>jar</type>
-	    <scope>provided</scope>
-	</dependency>
-	 
-	<dependency>
-	    <groupId>org.openmrs.web</groupId>
-	    <artifactId>openmrs-web</artifactId>
-	    <version>${openMRSVersion}</version>
-	    <type>test-jar</type>
-	    <scope>test</scope>
-	</dependency>
-	 
-	<dependency>
-	    <groupId>org.openmrs.test</groupId>
-	    <artifactId>openmrs-test</artifactId>
-	    <version>${openMRSVersion}</version>
-	    <type>pom</type>
-	    <scope>test</scope>
-	</dependency>
+
+		<dependency>
+			<groupId>org.openmrs.module</groupId>
+			<artifactId>namephonetics-api</artifactId>
+			<version>${project.parent.version}</version>
+		</dependency>
+
+		<dependency>
+			<groupId>org.openmrs.module</groupId>
+			<artifactId>legacyui-omod</artifactId>
+			<version>${legacyuiVersion}</version>
+			<scope>provided</scope>
+		</dependency>
+
+		<dependency>
+			<groupId>org.openmrs.web</groupId>
+			<artifactId>openmrs-web</artifactId>
+			<version>${openMRSVersion}</version>
+			<type>jar</type>
+			<scope>provided</scope>
+		</dependency>
+
+		<dependency>
+			<groupId>org.openmrs.web</groupId>
+			<artifactId>openmrs-web</artifactId>
+			<version>${openMRSVersion}</version>
+			<type>test-jar</type>
+			<scope>test</scope>
+		</dependency>
+
 	</dependencies>
 
 	<build>
@@ -167,7 +143,6 @@
 
 	<properties>
 		<MODULE_ID>${project.parent.artifactId}</MODULE_ID>
-		<MODULE_NAME>${project.parent.name}</MODULE_NAME>
 		<MODULE_VERSION>${project.parent.version}</MODULE_VERSION>
 		<MODULE_PACKAGE>${project.parent.groupId}.${project.parent.artifactId}</MODULE_PACKAGE>
 	</properties>
