+++ omod/src/main/resources/config.xml
@@ -2,7 +2,7 @@
 <!-- <!DOCTYPE module PUBLIC "-//OpenMRS//DTD OpenMRS Config 1.0//EN" "http://resources.openmrs.org/doctype/config-1.0.dtd">  -->
 <module configVersion="1.0">
 	<id>@MODULE_ID@</id>
-	<name>@MODULE_NAME@</name>
+	<name>Name Phonetics</name>
 	<version>@MODULE_VERSION@</version>
 	<package>@MODULE_PACKAGE@</package>
 	<author>PIH</author>
@@ -11,9 +11,8 @@
 	</description>
 
 	<activator>@MODULE_PACKAGE@.NamePhoneticsModuleActivator</activator>
-	
-	<!-- Require version 1.5 so jquery and jquery-ui are included. This could be relaxed -->
-	<require_version>1.6</require_version>
+
+	<require_version>${openMRSVersion}</require_version>
 	
 	<aware_of_modules>
 	    <aware_of_module>org.openmrs.module.legacyui</aware_of_module>
@@ -23,17 +22,6 @@
 		<point>org.openmrs.admin.list</point>
 		<class>@MODULE_PACKAGE@.web.extension.html.AdminList</class>
 	</extension>
-	
-	<conditionalResources>
-		<conditionalResource>
-			<path>/lib/namephonetics-api-1.*</path>
-			<openmrsVersion>1.*</openmrsVersion>
-		</conditionalResource>
-		<conditionalResource>
-			<path>/lib/namephonetics-api-2.*</path>
-			<openmrsVersion>2.*</openmrsVersion>
-		</conditionalResource>
-	</conditionalResources>
 	
 	<globalProperty>
 		<property>@MODULE_ID@.givenNameStringEncoder</property>
