+++ omod/src/main/resources/webModuleApplicationContext.xml
@@ -14,5 +14,4 @@
     <!-- for the controller -->
     <context:component-scan base-package="org.openmrs.module.namephonetics" />
 
-
 </beans>
