--- omod/src/test/java/org/openmrs/module/namephonetics/test/NamePhoneticsServiceTest.java
@@ -1,167 +1,0 @@
-package org.openmrs.module.namephonetics.test;
-
-import java.util.List;
-import java.util.Map;
-
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-import org.openmrs.Person;
-import org.openmrs.api.context.Context;
-import org.openmrs.module.namephonetics.NamePhonetic;
-import org.openmrs.module.namephonetics.NamePhoneticsService;
-import org.openmrs.web.test.BaseModuleWebContextSensitiveTest;
-import org.springframework.test.annotation.Rollback;
-
-
-public class NamePhoneticsServiceTest extends BaseModuleWebContextSensitiveTest {
-	
-    @Override
-    public Boolean useInMemoryDatabase(){
-        return true;
-    }
-    
-    
-	@Before
-	public void registerProcessers() {
-		NamePhoneticsService npService = Context.getService(NamePhoneticsService.class);	
-		npService.registerProcessor("Soundex", "org.apache.commons.codec.language.Soundex");
-	}
-        
-    @Test
-    @Rollback(false)
-    public void shouldReturnRenderers(){
-        NamePhoneticsService nps = Context.getService(NamePhoneticsService.class);
-        Map<String, String> map = nps.getProcessors();
-        for (Map.Entry<String, String> e : map.entrySet()){
-          System.out.println(e.getKey() + " " +e.getValue());
-        }
-        Assert.assertNotNull(map);
-        Assert.assertTrue(map.size() > 0);
-    }
-    
-    @Test
-    public void savePhoneticsForPerson_shouldSaveNamePhoneticsForPerson() {
-    	NamePhoneticsService npService = Context.getService(NamePhoneticsService.class);
-    	
-    	// create a name phonetics record for person #2 (Horatio Hornblower)
-    	Person person = Context.getPersonService().getPerson(2);
-    	npService.savePhoneticsForPerson(person, "Soundex", "Soundex", "Soundex", "Soundex");
-    	
-    	// get the newly-created Name Phonetics records and make sure they are correct
-    	List<NamePhonetic> results = npService.getNamePhoneticsByPersonName(person.getPersonName());
-  
-    	Assert.assertEquals(3, results.size());
-    	
-    	for (NamePhonetic n : results) {
-    		Assert.assertNotNull(n.getNameField());
-    		if (n.getNameField() == NamePhonetic.NameField.GIVEN_NAME) {
-    			Assert.assertEquals("H630", n.getRenderedString());
-    		}
-    		if (n.getNameField() == NamePhonetic.NameField.MIDDLE_NAME) {
-    			Assert.assertEquals("T230", n.getRenderedString());
-    		}
-    		if (n.getNameField() == NamePhonetic.NameField.FAMILY_NAME) {
-    			Assert.assertEquals("H651", n.getRenderedString());
-    		}
-    	}  	
-    }
-    
-    
-   
-    
-    
-//    @Test
-//    @Rollback(false)
-//    public void shouldShowAdditionalRenderer(){
-//        NamePhoneticsService nps = Context.getService(NamePhoneticsService.class);
-//        nps.registerProcessor("test", new Soundex());
-//        Map<String, StringEncoder> map = nps.getProcessors();
-//        Assert.assertTrue(map.size() == 6);
-//    }
-    
-//    @Test
-//    @Rollback(false)
-//    public void shouldEncodeSomeStrings(){
-//        NamePhoneticsService nps = Context.getService(NamePhoneticsService.class);
-//        System.out.println(nps.encodeString("Dave", "Chichewa Soundex"));
-//        System.out.println(nps.encodeString("Bluegrass", "Chichewa Soundex"));
-//        String st = nps.encodeString("MusicFFF", "Chichewa Soundex");
-//        System.out.println(st);
-//        Assert.assertTrue(st != null);
-//    }
-    
-//    
-//    @Test
-//    @Rollback(false)
-//    public void shouldSaveAndRetrieveANamePhonetic(){
-//        
-//        NamePhoneticsService nps = Context.getService(NamePhoneticsService.class);
-//        NamePhonetic np = new NamePhonetic();
-//        try {
-//        authenticate();
-//        Patient p = Context.getPatientService().getPatient(903);
-//        PersonName pn = p.getPersonName();
-//        Map<String, StringEncoder> map = nps.getProcessors();
-//        StringEncoder se = map.get("Caverphone");
-//        
-//        np.setNameField(NamePhonetic.NameField.MIDDLE_NAME);
-//        np.setPersonName(pn);
-//        np.setRenderedString(nps.encodeString(pn.getGivenName(), "Caverphone"));
-//        np.setRendererClassName(se.getClass().getName());
-//
-//        nps.saveNamePhonetic(np);
-//
-//        List<NamePhonetic>  list = nps.getNamePhoneticsByRenderedString(nps.encodeString(pn.getGivenName(), "Caverphone"), NamePhonetic.NameField.MIDDLE_NAME, se);
-//        System.out.println("size " + list.size());
-//
-//        NamePhonetic npnew = list.get(0);
-//        System.out.println("renderedString:" + npnew.getRenderedString() + " renderedClassName:" + npnew.getRendererClassName() + " nameField:" + npnew.getNameField() + " personName:" + npnew.getPersonName());
-//        
-//        } catch (Exception ex){
-//            System.out.println(ex.getMessage());
-//            ex.printStackTrace();
-//            Assert.assertFalse(true);
-//        }
-//
-//    }
-//    
-//    
-//    @Test
-//    @Rollback(false)
-//    public void shouldFindNames(){
-//    try {  
-//        authenticate();
-//        NamePhoneticsService nps = Context.getService(NamePhoneticsService.class);
-//        Patient p = Context.getPatientService().getPatient(905);
-//        nps.savePhoneticsForPatient(p);
-//        
-//        List<Patient> pList = nps.findPatient(p.getGivenName(), null, p.getFamilyName(), null);
-//        Assert.assertTrue(pList.size() > 0);
-//        for (Patient patient: pList){
-//            System.out.println("HERE's the patient " + patient + " " + patient.getGivenName() + " " + patient.getFamilyName() );
-//        }
-//    } catch (Exception ex){
-//        System.out.println(ex.getMessage());
-//        ex.printStackTrace();
-//        Assert.assertFalse(true);
-//    }
-//        
-//
-//    }
-    
-//    @Test
-//    @Rollback(false)
-//    public void testIdString(){
-//        String str = "13101,14865,14871,14872,14875,14878,14887,14889,14891,14904,14956,14968,14977,14997,15035,15036,15038,15039,15040,15041,15042,15043,15055,15058,15061,15065,15067,15068,15069,15095,15104,15118,15178,15182,15234,15235,15239,15254,15260,15274,15278,15281,15298,15312,15388,15393,15394,15523,15599,16248,16512,16543,16544,16545,16546,16547,16548,16549,16553,16554,16555,16556,16557,16558,16559,16560,16561,16562,16563,16564,16565,16566,16567,16568,16569,16570,16571,16572,16573,16574,16575,16576,16577,16581,16583,16584,16585,16586,16587,16588,16589,16590,16591,16592,16593,16594,16595,16596,16597,16598,,16600,16601,16602,16603,16606,16607,16608,16609,16610,16611,16612,16613,16614,16615,16616,16617,16618,16619,16620,16621,16622,16623,16624,16626,16628,16637,16639,16640,16641,16642,16648,16649,16650,16651,16652,16660,16661,16662,16663,16684,16694,16712,16726,16728,16729,16753,16757,16759,16793,16794,16795,16796,16805,16806,16807,16813,16814,16815,16816,16817,16818,16819,16821,16822,16826,16837,16838,16852,16855,16856,16857,16867,16868,16870,16871,16874,16876,16878,16879,16884,16887,16893,16908,16927,16940,16971,16974,16979,17007,17010,17012,17021,17024,17028,17043,17101,17123,17124,17125,17185,17186,17218,17229,17230,17229,17261,17264,17276,17351,17364,17366,17369,17376,17377,17378,17380,17381,17382,17383,17384,17389,17395,17405,17415,17416,17427,17428,17430,17431,17432,17440,17444,17445,17449,17455,17460,17467,17468,17469,17470,17480,17482,17484,17485,17486,17487,17488,17496,17497,17499,17500,17502,17503,17505,17506,17507,17513,17515,17522,17525,17526,17547,17560,17569,17587,17592,17623,17626,17629,17636,17640,17650,17652,17653,17654,17657,17659,17660,17661,17662,17664,17669,17671,17673,17675,17677,17681,17685,17695,17699,17721,17725,17730,17736,17740,17741,17742,17743,17745,17758,17789,17810,17811,17813,17817,17867,17871,17874,17875,17881,17889,17891,17892,17893,17894,17896,17901,17904,17907,17911,17912,17917,17920,17921,17926,17927,17929,17930,17931,17932,17937,17948,17954,17957,17958,17960,17964,17966,17969,17970,17976,17981,17985,17987,17989,17997,18001,18003,18004,18006,18010,18011,18012,18014,18015,18016,18018,18020,18021,18023,18025,18026,18029,18046,18072,18074,18075,18076,18077,18079,18080,18089,18092,18096,18102,18103,18104,18106,18107,18109,18111,18115,18116,18118,18119,18120,18121,18122,18133,18140,18141,18142,18143,18144,18145,18146,18150,18154,18169,18170,18175,18189,18190,18195,18198,18202,18203,18206,18208,18217,18218,18221,18222,18224,18230,18237,18240,18248,18280,18282,18287,18291,18299,18311,18312,18314,18317,18324,18326,18330,18333,18336,18337,18340,18345,18350,18366,18373,18380,18381,18398,18401,18413,18415,18434,18435,18448,18451,18452,18455,18462,18465,18467,18484,18485,18486,18502,18506,18519,18546,18556,18558,18560,18562,18570,18572,18575,18596,18599,18630,18651,18652,18655,18658,18660,18671,18678,18681,18685,18686,18687,18688,18689,18690,18691,18694,18695,18696,18697,18698,18700,18701,18702,18704,18706,18708,18711,18712,18713,18715,18717,18719,18720,18732,18735,18740,18741,18742,18751,18767,18770,18773,18777,18778,18785,18792,18793,18794,18797,18801,18802,18803,18804,18805,18806,18807,18808,18809,18812,18813,18814,18815,18816,18818,18819,18820,18822,18823,18825,18826,18828,18829,18830,18836,18843,18844,18847,18852,18860,18863,18866,18868,18870,18877,18878,18880,18882,18883,18885,18888,18893,18898,18901,18906,18917,18918,18919,18932,18933,18937,18941,18942,18944,18949,18950,18951,18954,18955,18958,18961,18963,18965,18967,18969,18971,18972,18974,18976,18977,18982,18984,18986,18987,18988,18989,18990,18993,18995,18997,18999,19000,19001,19002,19009,19015,19017,19017,19021,19022,19024,19025,19029,19030,19032,19036,19043,19044,19047,19052,19053,19055,19057,19065,19070,19073,19075,19081,19083,19084,19087,19095,19097,19098,19099,19101,19105,19106,19113,19117,19121,19125,19127,19131,19140,19143,19145,19156,19160,19161,19164,19166,19169,19172,19175,19176,19177,19178,19180,19182,19183,19184,19185,19197,19202,19204,19206,19218,19219,19220,19221,19222,19224,19225,19226,19227,19231,19232,19233,19234,19235,19237,19238,19239,19240,19242,19245,19248,19249,19250,19251,19252,19262,19264,19267,19269,19289,19291,19293,19294,19300,19301,19302,19303,19304,19305,19306,19308,19309,19310,19311,19313,19317,19318,19319,19321,19322,19326,19330,19334,19336,19337,19339,19340,19342,19343,19344,19345,19346,19347,19352,19353,19355,19358,19359,19360,19362,19363,19364,19368,19369,19370,19374,19375,19376,19377,19378,19383,19388,19391,19398,19402,19403,19405,19408,19414,19420,19424,19425,19427,19434,19436,19438,19439,19441,19443,19444,19445,19446,19449,19450,19451,19452,19453,19454,19461,19462,19474,19477,19480,19508,19510,19520,19522,19534,19535,19536,19539,19540,19541,19542,19544,19545,19546,19547,19548,19549,19550,19551,19553,19554,19555,19557,19558,19559,19560,19561,19562,19563,19564,19565,19566,19567,19568,19569,19570,19571,19573,19574,19575,19576,19577,19578,19579,19580,19581,19582,19583,19584,19585,19586,19587,19588,19589,19591,19592,19593,19594,19597,19598,19599,19604,19605,19607,19611,19612,19614,19615,19616,19617,19618,19619,19621,19622,19623,19626,19627,19628,19629,19630,19631,19632,19633,19634,19636,19637,19640,19642,19643,19644,19646,19647,19648,19649,19650,19651,19652,19653,19654,19656,19657,19658,19659,19660,19662,19663,19665,19666,19667,19668,19670,19671,19672,19673,19674,19675,19676,19677,19678,19680,19682,19684,19689,19690,19691,19692,19693,19694,19695,19696,19698,19699,19700,19701,19702,19708,19711,19712,19713,19718,19725,19727,19728,19729,19734,19736,19737,19742,19743,19744,19745,19746,19747,19748,19749,19750,19751,19752,19753,19754,19756,19757,19759,19761,19764,19765,19767,19768,19769,19770,19772,19773,19775,19776,19777,19781,19782,19783,19784,19785,19790,19792,19794,19798,19800,19801,19802,19804,19807,19808,19815,19817,19818,19819,19834,19840,19841,19843,19844,19845,19846,19847,19848,19850,19851,19853,19854,19855,19856,19858,19859,19860,19861,19862,19863,19864,19865,19866,19867,19868,19869,19870,19871,19872,19877,19878,19879,19882,19883,19884,19887,19888,19889,19891,19892,19893,19894,19896,19897,19898,19901,19902,19903,19904,19905,19906,19907,19909,19911,19912,19913,19914,19916,19917,19918,19919,19920,19921,19922,19923,19924,19926,19927,19928,19929,19932,19933,19934,19935,19936,19937,19938,19939,19941,19944,19949,19950,19951,19953,19954,19955,19957,19958,19960,19962,19963,19964,19965,19966,19967,19968,19969,19970,19971,19972,19973,19974,19975,19976,19977,19979,19980,19981,19982,19983,19984,19986,19987,19988,19990,19992,19993,19994,19995,19996,19997,19998,19999,20000,20001,20002,20003,20004,20006,20007,20008,20009,20010,20011,20012,20013,20014,20015,20017,20018,20019,20021,20022,20023,20025,20026,20027,20029,20030,20031,20032,20033,20034,20035,20036,20037,20038,20039,20040,20041,20042,20043,20044,20045,20046,20047,20048,20050,20051,20053,20056,20062,20064,20066,20067,20074,20077,20078,20080,20083,20086,20087,20088,20091,20097,20098,20101,20102,20104,20105,20106,20107,20108,20109,20110,20111,20112,20113,20114,20116,20117,20118,20119,20120,20121,20122,20123,20124,20126,20127,20128,20130,20131,20132,20133,20134,20136,20137,20140,20141,20142,20143,20144,20145,20146,20147,20148,20149,20150,20151,20152,20154,20155,20156,20158,20159,20160,20161,20162,20163,20164,20165,20166,20167,20168,20169,20170,20171,20172,20173,20174,20175,20198,20200,20201,20202,20203,20204,20205,20206,20207,20209,20210,20211,20212,20213,20214,20215,20216,20217,20218,20219,20221,20222,20223,20224,20225,20226,20227,20228,20229,20230,20231,20232,20233,20235,20243,20244,20245,20246,20281,20282,20283,20284,20285,20287,20288,20289,20291,20292,20293,20295,20296,20297,20298,20302,20303,20305,20306,20310,20312,20324,20325,20327,20328,20329,20330,20331,20332,20333,20335,20338,20339,20340,20341,20344,20345,20347,20348,20351,20352,20353,20354,20357,20358,20359,20360,20361,20362,20363,20364,20365,20366,20368,20369,20370,20371,20372,20373,20374,20375,20376,20377,20378,20380,20381,20382,20383,20384,20385,20386,20387,20388,20389,20390,20391,20393,20394,20396,20397,20398,20399,20400,20402,20404,20405,20407,20408,20413,20414,20415,20416,20417,20418,20419,20420,20422,20424,20425,20426,20428,20429,20431,20432,20433,20434,20435,20436,20437,20438,20439,20440,20443,20445,20446,20447,20448,20449,20450,20451,20452,20453,20454,20455,20456,20457,20458,20459,20460,20461,20462,20463,20466,20470,20473,20475,20476,20481,20483,20488,20493,20494,20495,20496,20498,20499,20500,20501,20502,20503,20504,20505,20506,20508,20511,20512,20513,20514,20515,20517,20522,20523,20524,20527,20528,20529,20530,20531,20532,20533,20534,20535,20536,20537,20538,20539,20540,20541,20542,20543,20544,20545,20546,20547,20548,20549,20550,20551,20552,20554,20555,20556,20558,20559,20560,20562,20563,20565,20566,20567,20568,20570,20571,20575,20577,20578,20579,20580,20581,20582,20584,20585,20586,20587,20588,20589,20590,20591,20592,20597,20598,20599,20600,20601,20602,20603,20604,20605,20606,20607,20608,20609,20610,20611,20612,20613,20617,20618,20619,20620,20621,20622,20623,20625,20626,20628,20631,20632,20637,20639,20644,20662,20665,20668,20670,20672,20673,20674,20676,20677,20681,20684,20685,20686,20687,20689,20690,20691,20692,20693,20694,20695,20696,20697,20700,20701,20702,20703,20705,20706,20708,20709,20710,20711,20712,20713,20714,20715,20716,20717,20718,20719,20720,20721,20722,20723,20724,20725,20726,20727,20729,20731,20733,20735,20736,20737,20738,20752,20756,20757,20758,20763,20764,20765,20766,20767,20769,20770,20771,20773,20774,20775,20776,20777,20779,20788,20790,20793,20794,20797,20798,20799,20800,20801,20803,20804,20805,20806,20807,20809,20810,20867,20868,20873,20874,20875,20891,20911,20912,20913,20914,20915,20916,20917,20918,20923,20940,20943,20968,20975,20978,20979,20980,20981,20989,20990,20993,20994,20998,21001,21006,21007,21008,21012,21013,21014,21015,21016,21017,21018,21020,21021,21022,21023,21024,21025,21026,21027,21028,21030,21031,21032,21033,21035,21036,21037,21038,21039,21040,21046,21049,21050,21052,21057,21058,21063,21065,21068,21069,21070,21071,21072,21073,21074,21075,21077,21079,21080,21081,21082,21084,21085,21086,21087,21088,21089,21090,21092,21123,21150,21161,21162,21163,21164,21165,21166,21168,21169,21171,21172,21175,21176,21190,21206,21225,21228,21230,21231,21232,21233,21234,21235,21236,21237,21239,21240,21242,21252,21259,21265,21270,21282,21290,21291,21295,21296,21318,21319,21321,21322,21323,21324,21325,21326,21327,21328,21329,21330,21331,21332,21333,21334,21335,21336,21337,21339,21341,21342,21345,21402,21404,21405,21406,21407,21408,21409,21410,21415,21416,21417,21418,21420,21421,21422,21423,21424,21425,21428,21429,21430,21431,21433,21435,21471,21474,21475,21476,21477,21478,21479,21480,21481,21482,21483,21484,21485,21486,21487,21488,21489,21490,21491,21493,21494,21495,21496,21497,21498,21499,21500,21501,21502,21503,21504,21505,21506,21507,21508,21509,21514,21515,21516,21522,21526,21528,21529,21532,21534,21535,21536,21537,21538,21542,21543,21544,21545,21546,21547,21548,21550,21551,21552,21554,21555,21558,21560,21562,21534,21628,21629,21630,21631,21632,21633,21634,21635,21636,21637,21638,21639,21640,21641,21642,21643,21644,21645,21646,21648,21649,21650,21651,21652,21653,21654,21656,21658,21659,21663,21664,21666,21667,21668,21669,21670,21676,21680,21681,21684,21685,21686,21687,21688,21689,21690,21692,21693,21704,21707,21709,21711,21713,21714,21715,21717,21718,21728,21730,21734,21745,21747,21748,21749,21750,21751,21752,21753,21754,21755,21756,21757,21758,21759,21760,21761,21762,21763,21764,21765,21766,21767,21768,21769,21770,21771,21772,21774,21776,21779,21780,21781,21782,21784,21787,21788,21790,21791,21793,21794,21795,21796,21797,21809,21810,21812,21813,21815,21816,21817,21818,21819,21820,21822,21825,21826,21827,21828,21829,21830,21831,21839,21840,21841,21842,21844,21845,21846,21847,21848,21849,21850,21851,21852,21853,21854,21855,21856,21857,21858,21859,21860,21862,21863,21864,21868,21869,21870,21874,21875,21878,21879,21880,21881,21882,21883,21884,21886,21887,21889,21890,21891,21892,21893,21894,21895,21901,21903,21904,21912,21915,21917,21919,21918,21922,21923,21924,21927,21928,21929,21930,21933,21936,21937,21938,21939,21940,21941,21942,21943,21944,21945,21947,21952,21954,21955,21965,21966,21967,21968,21969,21970,21971,21972,21973,21975,21984,21986,21987,22027,22028,22035,22037,22040,22041,22042,22046,22047,22049,22053,22054,22055,22056,22057,22060,22061,22062,22063,22064,22065,22068,22069,22070,22071,22072,22073,22074,22075,22076,22077,22078,22079,22080,22081,22082,22083,22084,22085,22088,22089,22090,22114,22127,22129,22131,22133,22153,22158,22159,22188,22198,22199,22200,22201,22202,22203,22204,22205,22206,22210,22212,22213,22214,22217,22218,22226,22231,22232,22233,22234,22235,22237,22238,22239,22240,22241,22242,22243,22244,22245,22246,22247,22248,22249,22250,22251,22252,22253,22255,22258,22261,22265,22267,22271,22272,22274,22275,22276,22279,22280,22291,22292,22293,22294,22295,22296,22297,22298,22301,22303,22305,22308,22309,22310,22311,22312,22314,22316,22317,22322,22323,22336,22339,22340,22342,22343,22344,22350,22351,22360,22366,22367,22373,22374,22375,22376,22377,22381,22383,22386,22390,22391,22393,22394,22395,22396,22397,22401,22403,22404,22405,22406,22407,22409,22410,22411,22412,22414,22416,22417,22418,22419,22420,22421,22422,22423,22424,22425,22426,22427,22428,22429,22430,22432,22433,22434,22435,22436,22437,22438,22439,22440,22441,22443,22444,22446,22447,22457,22459,22460,22461,22462,22463,22464,22465,22466,22467,22479,22481,22482,22484,22485,22486,22488,22490,22493,22494,22495,22496,22497,22498,22499,22500,22503,22504,22505,22507,22508,22511,22512,22516,22517,22518,22519,22520,22521,22523,22525,22527,22529,22530,22534,22535,22536,22537,22538,22539,22540,22541,22542,22543,22544,22545,22547,22551,22553,22554,22558,22564,22578,22579,22580,22581,22582,22585,22591,22592,22593,22594,22596,22597,22598,22599,22601,22602,22608,22611,22612,22614,22616,22617,22620,22622,22623,22624,22625,22626,22628,22633,22637,22638,22640,22641,22642,22643,22646,22647,22648,22649,22650,22651,22652,22653,22654,22655,22656,22657,22658,22659,22660,22661,22663,22667,22668,22669,22670,22671,22676,22679,22682,22684,22685,22686,22687,22688,22689,22690,22691,22692,22693,22694,22695,22696,22697,22698,22700,22701,22719,22722,22724,22725,22726,22727,22728,22729,22730,22732,22733,22734,22735,22736,22737,22738,22739,22741,22742,22743,22744,22745,22746,22750,22751,22752,22753,22754,22755,22758,22759,22760,22761,22762,22768,22771,22772,22773,22774,22775,22776,22779,22788,22789,22790,22793,22794,22795,22796,22797,22798,22799,22801,22802,22803,22804,22805,22807,22808,22809,22810,22812,22813,22814,22815,22816,22819,22820,22821,22822,22823,22824,22826,22827,22830,22831,22832,22833,22834,22837,22838,22839,22847,22849,22853,22854,22855,22856,22857,22858,22859,22860,22862,22863,22864,22865,22866,22867,22868,22869,22870,22871,22872,22873,22876,22878,22882,22883,22884,22889,22890,22891,22892,22894,22895,22904,22905,22906,22907,22908,22909,22912,22913,22917,22919,22920,22921,22922,22923,22924,22925,22926,22927,22928,22929,22930,22931,22932,22933,22934,22935,22937,22938,22939,22940,22941,22942,22944,22945,22946,22947,22948,22950,22951,22952,22953,22954,22955,22956,22957,22958,22959,22960,22961,22962,22963,22964,22965,22967,22968,22969,22970,22971,22972,22973,22974,22975,22976,22978,22979,22980,22981,22982,22983,22984,22985,22986,22987,22989,22993,22995,22999,23001,23002,23004,23005,23008,23009,23012,23013,23014,23015,23016,23017,23018,23019,23020,23022,23023,23024,23025,23026,23027,23028,23029,23030,23031,23022,23023,23024,23025,23026,23027,23028,23029,23030,23031,23035,23036,23037,23038,23042,23043,23045,23046,23047,23048,23049,23050,23051,23053,23054,23055,23056,23057,23058,23059,23060,23061,23063,23065,23066,23068,23069,23070,23071,23072,23073,23074,23075,23076,23079,23088,23090,23092,23093,23094,23095,23096,23097,23098,23099,23100,23101,23102,23103,23104,23105,23106,23107,23108,23109,23110,23111,23112,23113,23114,23115,23117,23118,23119,23120,23123,23124,23125,23126,23127,23128,23129,23130,23131,23132,23133,23134,23135,23136,23137,23140,23141,23142,23143,23144,23145,23146,23148,23149,23150,23151,23153,23154,23155,23159,23160,23165,23166,23169,23195,23201,23205,23221,23239,23297,23302,23309,23311,23312,23323,23324,23337,23347,23348,23349,23350,23351,23352,23353,23354,23355,23356,23358,23360,23361,23380,23383,23384,23413,23423,23424,23425,23426,23427,23428,23436,23438,23444,23446,23488,23502,23503,23504,23506,23507,23508,23548,23549,23558,23560,23561,23592,23595,23596,23610,23634,23635,23636,23637,23638,23644,23645,23646,23647,23651,23653,23662,23664,23669,23671,23673,23675,23679,23680,23681,23685,23775,23776,23777,23782,23797,23800,23815,23817,23819,23821,23824,23826,23827,23828,23829,23847,23850,23854,23856,23857,23866,23879,23882,23885,23893,23896,23897,23898,23899,23900,23901,23902,23907,23913,23914,23915,23917,23919,23921,23922,23931,23932,23933,23934,23935,23936,23938,23941,23942,23943,23945,23946,23948,23949,23950,23951,23952,23953,23954,23955,23956,23957,23963,23966,23972,23975,23976,23985,23987,23988,23990,23994,23997,23999,24000,24001,24006,24014,24023,24024,24029,24031,24032,24047,24058,24070,24073,24082,24084,24089,24090,24092,24097,24102,24110,24122,24123,24136,24144,24148,24164,24165,24166,24167,24168,24196,24199,24207,24208,24209,24212,24217,24227,24236,24238,24239,24241,24243,24254,24260,24280,24281,24285,24295,24296,24305,24310,24313,24323,24332,24334,24341,24343,24356,24359,24360,24361,24375,24376,24377,24378,24380,24381,24383,24384,24387,24390,24399,24400,24401,24405,24414,24415,24422,24441,24471,24472,24475,24478,24479,24485,24496,24505,24518,24532,24534,24539,24543,24563,24587,24588,24589,24590,24595,24596,24599,24600,24603,24607,24616,24617,24618,24619,24624,24625,24627,24628,24630,24636,24641,24648,24655,24659,24663,24664,24672,24674,24675,24684,24699,24700,24701,24704,24707,24708,24709,24729,24730,24732,24735,24747,24758,24767,24773,24774,24776,24790,24799,24820,24830,24846,24863,24864,24865,24867,24869,24870,24871,24873,24874,24878,24879,24880,24881,24882,24883,24885,24886,24887,24888,24889,24890,24892,24901,24902,24903,24913,24915,24926,24927,24929,24937,24945,24947,24948,24949,24950,24951,24952,24956,24963,24964,24968,24971,24977,24980,24990,24993,24994,24996,24997,24998,24999,25000,25001,25002,25003,25004,25005,25006,25007,25008,25009,25010,25012,25013,25015,25016,25017,25018,25019,25022,25034,25036,25037,25038,25042,25052,25053,25054,25057,25058,25060,25066,25075,25078,25079,25084,25086,25087,25088,25089,25091,25092,25093,25094,25096,25097,25098,25100,25101,25102,25103,25104,25107,25108,25109,25110,25111,25112,25114,25119,25132,25133,25134,25147,25148,25149,25150,25152,25154,25155,25156,25158,25160,25163,25165,25167,25185,25191,25211,25214,25216,25227,25228,25237,25239,25240,25242,25243,25244,25246,25267,25272,25280,25319,25320,25321,25322,25323,25325,25326,25328,25330,25331,25334,25335,25337,25340,25341,25359,25361,25363,25397,25400,25402,25403,25404,25405,25409,25415,25425,25437,25449,25450,25455,25456,25460,25462,25464,25469,25472,25473,25488,25489,25490,25491,25492,25391,25507,25511,25539,25543,25544,25545,25546,25547,25552,25555,25560,25561,25562,25565,25588,25589,25590,25591,25592,25596,25610,25615,25616,25617,25635,25640,25650,25663,25672,25683,25737,25738,25746,25750,25751,25752,25753,25754,25768,25811,25822,25823,25829,25848,25849,25891,25896,25904,25914,25922,25925,25932,25933,25934,25936,25953,25954,25965,25976,25981,25982,26012,26071,26073,26075,26085,26087,26100,26101,26112,26120,26134,26148,26157,26160,26166,26171,26172,26173,26174,26175,26176,26177,26178,26179,26180,26181,26217,26219,26224,26225,26230,26232,26233,26245,26258,26287,26293,26297,26323,26350,26351,26358,26362,26372,26403,26405,26431,26440,26451,26452,26459,26460,26461,26472,26475,26476,26478,26506,26510,26566,26578,26579,26580,26582,26583,26584,26585,26637,26653,26654,26656,26657,26670,26671,26672,26676,26681,26686,26687,26688,26689,26729,26739,26746";
-//        Set<Integer> ids = new HashSet<Integer>();
-//        for (StringTokenizer st = new StringTokenizer(str, ","); st.hasMoreTokens(); ) {
-//            String s = st.nextToken().trim();
-//            ids.add(Integer.valueOf(s.trim()));
-//        }
-//        System.out.println("Set size " + ids.size());
-//        Assert.assertTrue(true);
-//    }
-    
-    
-}
