--- omod/src/test/resources/test-hibernate.cfg.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!DOCTYPE hibernate-configuration PUBLIC
-	"-//Hibernate/Hibernate Configuration DTD 3.0//EN"
-	"http://hibernate.sourceforge.net/hibernate-configuration-3.0.dtd">
-
-<hibernate-configuration>
-	<session-factory>
-        <mapping resource="NamePhonetic.hbm.xml" />
-	</session-factory>
-</hibernate-configuration>
